Created helm template for secret checks
These checks run on pre-install/pre-upgrade hooks (depending on nature
of helm release).

The only weirdness so far is that each key also requires a displayName.
This is the name for the container in which that key is checked with a
`stat`. The reason for this is that some secret keys (e.g.
`"sacred.json"`) are not valid container names (in that case, because of
the `.`). Rather than try to solve the problem with a helm template
function, this is more reliable even if it places a little more burden
on the user.