cloudbuild_build_deploy deploys the right images (#141)
Fixes https://github.com/doc-ai/tensorio-models/issues/139

This also required some changes to the helm templates (and the logic
around rendering different templates). There are minimal changs to the
fixtures (just whitespace), but the templates and sample values files
have been updated to reflect these changes.

It is now possible to set `release` flags on each component to white- or
black-list it from a release.